클래스를 나누는 리팩토링 수행
